Split all collective unit test files into separate files
